Move inclusions to .hpp file
